PYTHON[ COMPLETE DIAMOND PROBLEM]
This is a method I felt easy and understandable. 
We use simple loops, brute force(I assume they call it) to create spaces or symbols of your liking under charval and another value spaceval which helps to print the word[string str] and later in the same print we send the symbols again  which forms a uniform diamond.